% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_url.R
\name{append_url}
\alias{append_url}
\title{Construct a full SQUIDLE API URL}
\usage{
append_url(api, endpoint, query_filters = NULL, query_parameters = NULL)
}
\arguments{
\item{api}{An instance of SQAPI, containing properties \code{"host"} and \code{"auth"}.}

\item{endpoint}{A character string specifying the endpoint. See
"https://squidle.org/api/help?template=api_help_page.html" for endpoint details.}

\item{query_filters}{The output from \code{SQAPI::query_filter()}. A named list of filters.
These are appended to the url inside of the \code{q={}} json string.}

\item{query_parameters}{The output from \code{SQAPI::query_params()}. A list of two elements:
\itemize{
\item{\code{q}} {The first element is a list of JSON-encoded query parameters (limit, offset, order_by, group_by, single).
These parameters will exist within the \code{q={}} json string in the final url.}
\item{\code{qparams}} {The second element is a list of top-level query parameters (template, disposition, include_columns, page, results_per_page).
These parameters will exist outside of the \code{q={}} json string in the final url.}
}}
}
\value{
A character string representing the full URL encoded to meet the SQUIDLE API requirements.
}
\description{
Constructs a complete URL by appending an endpoint to the API host,
optionally adds query filters and query parameters. Filters
and parameters are passed to this function directly from the output of \code{SQAPI::query_filter()}
and \code{SQAPI::query_params}. Typically used as a helper function in \code{SQAPI::request()} and
\code{SQAPI::export()}
}
