% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fragstats.R
\name{count_border_vertices}
\alias{count_border_vertices}
\title{Count how many (unclosed) polygon vertices lie on the image border}
\usage{
count_border_vertices(
  coords,
  media_width,
  media_height,
  edge_tolerance = 1e-06
)
}
\description{
Count how many (unclosed) polygon vertices lie on the image border
}
\keyword{internal}
