% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_media}
\alias{dummy_media}
\title{Dummy Media Data}
\format{
\subsection{\code{dummy_media}}{

A data frame with 131 obs. and 1 variable:
\describe{
\item{point.media.key}{point media key}
}
}
}
\usage{
dummy_media
}
\description{
A subset of data from SQUIDLE that contains "point.media.key" that relate to a set of images.
Used in POST workflow example in README.
}
\keyword{datasets}
