% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_key}
\alias{find_key}
\title{Search a List and Return a Value}
\usage{
find_key(lst, key)
}
\arguments{
\item{lst}{A list to search. Can be nested.}

\item{key}{A character string representing the key to search for.}
}
\value{
The value associated with the first occurrence of the key, or \code{NULL} if the key is not found.
}
\description{
Internal helper to recursively search a nested list for a key and return the associated value.
}
