% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_filetype}
\alias{get_filetype}
\title{Search for file type in response headers}
\usage{
get_filetype(response)
}
\arguments{
\item{response}{Response object from a request or export call.}
}
\value{
String representing file type.
}
\description{
Internal helper that searches for file type in response headers.
}
