% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{make_export_request}
\alias{make_export_request}
\title{Helper to Make an HTTP Request to the SQUIDLE API Export Endpoint}
\usage{
make_export_request(url, token, write_to_disk = FALSE, file = NULL)
}
\arguments{
\item{url}{A character string representing the full URL for the API export request.}

\item{token}{A character string containing the authentication token to include in the request header.}

\item{write_to_disk}{Logical; if \code{TRUE}, the response will be written to disk.
If \code{FALSE}, the response will be stored in memory.}

\item{file}{A character string specifying the filename to save the response to when \code{write_to_disk = TRUE}.
This argument is ignored if \code{write_to_disk = FALSE}.}
}
\value{
An \code{httr::response} object. If \code{write_to_disk = TRUE}, the response is written to disk.
If \code{write_to_disk = FALSE}, the response content is stored in memory.
}
\description{
Internal helper that sends a GET request to the SQUIDLE API export endpoint. It can either
write the result to disk or return the content in memory, based on the provided arguments.
}
\details{
This function is used internally within other functions (specifically export) to make API requests. It is flexible in how it handles
the response, allowing either a file save to disk or in-memory storage for further processing.
}
\keyword{internal}
