% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{make_request}
\alias{make_request}
\title{Helper to Make an HTTP Request to the SQUIDLE API Endpoint}
\usage{
make_request(verb, url, token, body = NULL)
}
\arguments{
\item{verb}{A valid http verb.}

\item{url}{A character string representing the full URL for the API export request.}

\item{token}{A character string containing the authentication token to include in the request header.}

\item{body}{A named R list to include as the JSON body of the request (for POST or PATCH). Can be \code{NULL}.}
}
\value{
An \code{httr::response} object.
}
\description{
Internal helper that sends a GET request to the SQUIDLE API endpoint. Adds headers as needed depending on verb.
}
\details{
This function is used internally within request to make API requests.
}
\keyword{internal}
