% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{poll_for_result}
\alias{poll_for_result}
\title{Helper for Status polling logic for SQUIDLE export}
\usage{
poll_for_result(api, json, write_disk, filename)
}
\arguments{
\item{api}{An instance of SQAPI, containing properties \code{"host"} and \code{"auth"}.}

\item{json}{The parsed JSON response from the initial export request.}

\item{write_disk}{Logical; whether to write the result to disk.}

\item{filename}{Filename to save the result if write_disk is TRUE.}
}
\value{
An httr::response object.
}
\description{
Internal helper that polls a status URL until the result is ready,
then retrieves the final result from the result URL.
}
