% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_filter.R
\name{query_filter}
\alias{query_filter}
\title{Create a Query Filter for SQUIDLE API Requests}
\usage{
query_filter(name, op, val = NULL)
}
\arguments{
\item{name}{A character string specifying the name of the field to filter.}

\item{op}{A character string specifying the operator (e.g., \code{"=="}, \code{"!="}, \code{"in"}).}

\item{val}{The value to compare the field against. Can be another filter (i.e., nested).}
}
\value{
A named list representing the filter, to be included in your query.
}
\description{
Constructs a named list representing a query filter. The filter includes
a name, an operator , and a value.
Can be used to created nested lists for nested queries. Details on fields and operators
can be found at "https://squidle.org/api/help?template=api_help_page.html"
}
\examples{
query_filter("id", "eq", "5432")

query_filter(
name = "media",
op = "any",
val = query_filter(
  name = "deployment",
  op = "has",
  val = query_filter(
   name = "campaign",
   op = "has",
   val = query_filter(name = "key", op = "eq", val = "Batemans201011")
  )
)
)

}
