% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQAPI.R
\name{SQAPI}
\alias{SQAPI}
\title{Create class "SQAPI" with property "host" and "auth"}
\description{
Class SQAPI has a host and authorisation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{host}}{Host url - defaults to "https://squidle.org".}

\item{\code{auth}}{Authorisation for your SQUIDLE API account (your API token).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SQAPI-new}{\code{SQAPI$new()}}
\item \href{#method-SQAPI-clone}{\code{SQAPI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQAPI-new"></a>}}
\if{latex}{\out{\hypertarget{method-SQAPI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of SQAPI that has a host and authorisation.This will handle authorisation of all request to SQUIDLE API in an "r" session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQAPI$new(host = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{Host url - defaults to "https://squidle.org".}

\item{\code{auth}}{Your SQUIDLE API token.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SQAPI} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQAPI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SQAPI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQAPI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
