% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sqapi_transform}
\alias{sqapi_transform}
\title{Skeleton function to build a transform object for API queries}
\usage{
sqapi_transform(...)
}
\arguments{
\item{...}{One or more lists, each representing an operation.}
}
\value{
A list of operations suitable for \code{SQAPI::request} or \code{SQAPI::export}
}
\description{
Constructs a \code{transform} list object to pass to the SQUIDLE API request or export function.
}
\examples{
# Single operation
transform(list(module = "pandas", method = "json_normalize"))

}
