% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_url.R
\name{append_url}
\alias{append_url}
\title{Construct a full SQUIDLE API URL}
\usage{
append_url(
  api,
  endpoint,
  query_filters = NULL,
  query_parameters = NULL,
  template = NULL,
  disposition = NULL,
  transform = FALSE
)
}
\arguments{
\item{api}{An instance of \code{SQAPI}, containing properties \code{"host"} and \code{"auth"}.}

\item{endpoint}{A character string specifying the endpoint. See
"https://squidle.org/api/help?template=api_help_page.html" for endpoint details.}

\item{query_filters}{The output from \code{SQAPI::query_filter()}. A named list of filters.
These are appended to the url inside of the \code{q={}} json string.}

\item{query_parameters}{The output from \code{SQAPI::query_params()}. A list of two elements:
\itemize{
\item{\code{q}} {A list of query parameters (e.g., \code{limit}, \code{offset}, \code{order_by}, \code{group_by}, \code{single})
to be encoded as JSON within the \code{q={}} string.}
\item{\code{qparams}} {A list of top-level query parameters (e.g., \code{include_columns}, \code{page}, \code{results_per_page})
that appear outside the \code{q={}} string.}
}}

\item{template}{Optional character string to specify the template to use (e.g., "data.csv"). Default is JSON.}

\item{disposition}{Optional character string to specify content disposition. Accepts \code{"attachment"} and \code{"inline"}.}
}
\value{
A character string representing the full URL encoded to meet the SQUIDLE API requirements.
}
\description{
Constructs a complete URL by appending an endpoint to the API host,
optionally adds query filters and query parameters. Filters
and parameters are passed to this function directly from the output of \code{SQAPI::query_filter()}
and \code{SQAPI::query_params}. Typically used as a helper function in \code{SQAPI::request()} and
\code{SQAPI::export()}
}
